requirements.txt
Flask==1.0.2
Flask-Mail==0.9.1
Flask-WTF==0.14.2
gunicorn==19.9.0
python-dotenv==0.10.1